Create Declarations and Array Access:
ariable Declarations and Array Access:
Declares variables for persons and an array. Prints elements of the array.
Variable Name Change and Array Access:
Renames variables and accesses array elements.
Array of Persons with Details and Loop:
Defines an array of persons with details and prints each person's information using a loop.
Loop and Conditional Statement (Even Ages):
Iterates through an array of ages, prints even ages.
Array of Users and Loop with Conditional Statement:
Defines an array of users, prints first names if the gender is "female."
Simple Function:
Defines a function to sum two numbers. Calls the function with different arguments and prints the results.